Add files via upload
This is a card game "simulation" played by a a user-given number of players with results essentially determined by a random number generator. The user also inputs the size of each players deck and the turn limit on the game. More information can be found in the README.txt found inside the zip file.